travel_tourism_hospitality/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
cachetools==4.2.2
certifi==2022.12.7
chartjs==1.2
Django==4.1.3
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
telegram==0.0.1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
